requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
black==25.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
datasets==3.5.0
dill==0.3.8
docopt-ng==0.9.0
farama-notifications==0.0.4
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.4.0
gitdb==4.0.12
gitpython==3.1.44
google-auth==2.39.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.2.0
grpcio==1.71.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.13.0
huggingface-hub==0.30.2
huggingface-sb3==3.0
idna==3.10
imitation==1.0.1
importlib-metadata==8.6.1
importlib-resources==6.5.2
jax==0.4.30
jax-jumpy==1.0.0
jaxlib==0.4.30
jinja2==3.1.6
joblib==1.4.2
jsonpickle==4.0.5
keras==3.9.2
kiwisolver==1.4.7
libclang==18.1.1
mako==1.3.10
markdown==3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
munch==4.0.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
opencv-python==4.11.0.86
opt-einsum==3.4.0
optree==0.15.0
optuna==4.3.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
pipdeptree==2.26.0
platformdirs==4.3.7
propcache==0.3.1
protobuf==4.25.6
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pygame==2.6.1
pyglet==1.5.31
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9
sacred==0.8.7
sb3-contrib==2.2.1
scikit-learn==1.6.1
scipy==1.13.1
seals==0.2.1
setuptools==68.2.2
shimmy==1.3.0
six==1.17.0
smmap==5.0.2
sqlalchemy==2.0.40
stable-baselines3==2.2.1
stable-retro==0.9.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
werkzeug==3.1.3
wrapt==1.14.1
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0